+++ src/tkc/slist.h
@@ -97,6 +97,12 @@ typedef struct _slist_t {
    * 节点内存分配器。
    */
   mem_allocator_t* node_allocator;
+
+  /**
+   * @property {mem_allocator_t*} node_allocator_is_shared
+   * 节点内存分配器是否共享。
+   */
+  bool_t node_allocator_is_shared;
 } slist_t;
 
 /**
@@ -304,6 +310,16 @@ ret_t slist_reverse(slist_t* slist);
  * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
  */
 ret_t slist_set_node_allocator(slist_t* slist, mem_allocator_t* allocator);
+
+/**
+ * @method slist_set_shared_node_allocator
+ * 设置共享节点内存分配器。
+ * @param {slist_t*} slist 单向链表对象。
+ * @param {mem_allocator_t*} allocator 内存分配器对象。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t slist_set_shared_node_allocator(slist_t* slist, mem_allocator_t* allocator);
 
 /**
  * @method slist_deinit